refactor: extract useRoomLifecycle hook from RoomView
